+++ Templates/Scripts/Runtime/Function/T_Func+Global.cs
@@ -13,7 +13,7 @@ using UnityEngine.Purchasing;
 public static partial class Func {
 	#region 클래스 함수
 	/** 아이템을 획득한다 */
-	public static void AcquireItem(STItemInfo a_stItemInfo, int a_nExtraNumItems = KCDefine.B_VAL_0_INT) {
+	public static void AcquireItem(STItemInfo a_stItemInfo, long a_nExtraNumItems = KCDefine.B_VAL_0_INT) {
 		switch(a_stItemInfo.m_eItemKinds) {
 			case EItemKinds.GOODS_COINS: {
 				CUserInfoStorage.Inst.AddNumCoins(a_stItemInfo.m_nNumItems + a_nExtraNumItems);
@@ -41,7 +41,7 @@ public static partial class Func {
 	}
 
 	/** 아이템을 구입한다 */
-	public static void BuyItem(STSaleItemInfo a_stSaleItemInfo, List<int> a_oExtraNumItemsList = null, int a_nExtraPrice = KCDefine.B_VAL_0_INT, bool a_bIsIgnoreAcquire = false) {
+	public static void BuyItem(STSaleItemInfo a_stSaleItemInfo, List<long> a_oExtraNumItemsList = null, long a_nExtraPrice = KCDefine.B_VAL_0_INT, bool a_bIsIgnoreAcquire = false) {
 		// 아이템 획득이 가능 할 경우
 		if(!a_bIsIgnoreAcquire) {
 			for(int i = 0; i < a_stSaleItemInfo.m_oItemInfoList.Count; ++i) {
